+++ VeryCoolEngine/src/VeryCoolEngine/Input.h
@@ -7,11 +7,11 @@ namespace VeryCoolEngine {
 	public:
 		static bool IsKeyPressed(int keyCode) { return _sInstance->IsKeyPressedImpl(keyCode); }
 		static bool IsMouseButtonPressed(int button) { return _sInstance->IsMouseButtonPressedImpl(button); }
-		static std::pair<float, float> GetMousePos() { return _sInstance->GetMousePosImpl(); }
+		static std::pair<double, double> GetMousePos() { return _sInstance->GetMousePosImpl(); }
 	protected:
 		virtual bool IsKeyPressedImpl(int keyCode) = 0;
 		virtual bool IsMouseButtonPressedImpl(int button) = 0;
-		virtual std::pair<float, float> GetMousePosImpl() = 0;
+		virtual std::pair<double, double> GetMousePosImpl() = 0;
 	private:
 		static Input* _sInstance;
 	};
